+++ src/pages/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Head from 'next/head';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
 import { ILocale } from '@/interfaces/locale';
-import Header from '@/components/common/header';
+import Header from '@/components/common/header_1';
 import HomePageBody from '@/components/home_page/home_page_body';
 import Footer from '@/components/common/footer';
